Refactor reward calculation into separate function
